requirements.txt
transformers==4.37.0
accelerate==0.21.0
scipy==1.11.3
bitsandbytes>=0.41.0
datasets>=2.14.5
ogb==1.2.0
sentencepiece>=0.1.99
peft>=0.4.0
deepspeed==0.15.4
torch_geometric
rdkit>=2021.03.5
selfies>=1.0.0
trl==0.4.7
wandb
scikit-learn
nltk
Levenshtein
rouge-score
pytest
flask